feat: :sparkles: added refactor
